requirements.txt
Flask==1.1.1
flask-cors==3.0.8
python-dotenv==0.10.3
gunicorn==19.9.0
google-cloud-storage==1.23.0
google-cloud-tasks==1.3.0
googleapis-common-protos==1.6.0
# torch
https://download.pytorch.org/whl/cpu/torch-1.0.1.post2-cp36-cp36m-linux_x86_64.whl
torchvision==0.4.1